DABOM-229 refactor: service 인터페이스 적용 및 Helper 클래스 디렉토리 이동
