+++ src/main/java/com/project/domain/usage/service/helper/CustomerQuotaWriter.java
@@ -1,4 +1,4 @@
-package com.project.domain.usage.service;
+package com.project.domain.usage.service.helper;
 
 import java.time.LocalDate;
