+++ src/main/java/com/project/domain/usage/service/helper/UsageEventValidator.java
@@ -1,4 +1,4 @@
-package com.project.domain.usage.service;
+package com.project.domain.usage.service.helper;
 
 import org.springframework.stereotype.Component;
