+++ src/test/java/com/project/domain/usage/infra/messaging/UsageEventsConsumerTest.java
@@ -21,8 +21,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.project.domain.usage.service.UsageEventValidator;
-import com.project.domain.usage.service.UsageSyncService;
+import com.project.domain.usage.service.UsageSyncServiceImpl;
+import com.project.domain.usage.service.helper.UsageEventValidator;
 import com.project.global.event.dto.EventEnvelope;
 import com.project.global.event.dto.usage.UsagePayload;
 
@@ -33,7 +33,7 @@ class UsageEventsConsumerTest {
 
     @Mock private ObjectMapper objectMapper;
 
-    @Mock private UsageSyncService usageSyncService;
+    @Mock private UsageSyncServiceImpl usageSyncServiceImpl;
 
     @Mock private UsageEventValidator validator;
 
@@ -57,7 +57,7 @@ class UsageEventsConsumerTest {
         consumer.consume(consumerRecord);
 
         // then
-        verify(usageSyncService, times(1))
+        verify(usageSyncServiceImpl, times(1))
                 .syncUsage(
                         eq(envelope.eventId()),
                         anyString(), // timestamp string
@@ -84,7 +84,7 @@ class UsageEventsConsumerTest {
         consumer.consume(consumerRecord);
 
         // then
-        verify(usageSyncService, never()).syncUsage(any(), any(), any());
+        verify(usageSyncServiceImpl, never()).syncUsage(any(), any(), any());
     }
 
     @Test
@@ -102,6 +102,6 @@ class UsageEventsConsumerTest {
         consumer.consume(consumerRecord);
 
         // then
-        verify(usageSyncService, never()).syncUsage(any(), any(), any());
+        verify(usageSyncServiceImpl, never()).syncUsage(any(), any(), any());
     }
 }
