+++ src/test/java/com/project/domain/usage/service/UsageEventValidatorTest.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import com.project.domain.usage.service.helper.UsageEventValidator;
 import com.project.global.event.dto.usage.UsagePayload;
 
 class UsageEventValidatorTest {
